% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_calc.R
\docType{data}
\name{base}
\alias{base}
\title{tp_calc}
\format{An object of class \code{data.frame} with 4 rows and 2 columns.}
\usage{
base
}
\arguments{
\item{base}{A data frame generated by function base_sig}

\item{fish}{A data frame with isotope data arranged by species and site}

\item{frac}{Trophic fractionation factor, default set at 3.4}
}
\value{
A dataframe of columns sitename, d13C, d15N, species, id, and tp
}
\description{
Takes output from base_sig to calculate the trophic position for each fish (row) in dataframe
}
\seealso{
\code{\link{base_sig}}

\code{\link{fcl_calc}}

\code{\link{plot_web}}
}
\keyword{datasets}
