% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sig.R
\name{base_sig}
\alias{base_sig}
\title{base_sig}
\usage{
base_sig(dat, frac = 3.4)
}
\arguments{
\item{dat}{A data frame with primary consumer (mussel) d15N data}

\item{frac}{Trophic fractionation factor, default set at 3.4}
}
\value{
A dataframe of columns sitename and baseline
}
\description{
Takes primary producer d15N signature data separated by site and returns a dataframe with two columns
}
