% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcl_calc.R
\name{fcl_calc}
\alias{fcl_calc}
\title{fcl_calc}
\usage{
fcl_calc(tpdf)
}
\arguments{
\item{tp}{A data frame generated by function tp_calc}
}
\value{
A dataframe of columns sitename and fcl
}
\description{
Takes output from tp_calc to calculate the maximum trophic position (food chain length) for each site in dataframe
}
