% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sig.R, R/tp_calc.R
\docType{data}
\name{mussurl}
\alias{mussurl}
\alias{mussurl}
\title{base_sig}
\format{An object of class \code{character} of length 1.}
\usage{
mussurl

mussurl
}
\arguments{
\item{muss}{A data frame with primary consumer (mussel) d15N data}

\item{frac}{Trophic fractionation factor, default set at 3.4}

\item{base}{A data frame generated by function base_sig}

\item{fish}{A data frame with isotope data arranged by species and site}

\item{frac}{Trophic fractionation factor, default set at 3.4}
}
\value{
A dataframe of columns sitename and baseline

A dataframe of columns sitename, d13C, d15N, species, id, and tp
}
\description{
Takes primary producer d15N signature data separated by site and returns a dataframe with two columns

Takes output from base_sig to calculate the trophic position for each fish (row) in dataframe
}
\seealso{
\code{\link{base_sig}}

\code{\link{fcl_calc}}

\code{\link{plot_web}}
}
\keyword{datasets}
