% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_web.R
\name{plot_web}
\alias{plot_web}
\title{plot_web}
\usage{
plot_web(fish)
}
\arguments{
\item{fish}{A dataframe with d13C and d15N data for fish at different sites along a river}
}
\value{
A plot in isotopic space (x=d13C, y=d15N)
}
\description{
Generates a plot in isotopic space for each study site
}
