% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tp_calc.R
\name{tp_calc}
\alias{tp_calc}
\title{tp_calc}
\usage{
tp_calc(base, dat, frac = 3.4)
}
\arguments{
\item{base}{A data frame generated by function base_sig}

\item{dat}{A data frame with isotope data arranged by species and site}

\item{frac}{Trophic fractionation factor, default set at 3.4}
}
\value{
A dataframe of columns sitename, d13C, d15N, species, id, and tp
}
\description{
Takes output from base_sig to calculate the trophic position for each fish (row) in dataframe
}
\examples{
tp_calc(base = iris, dat = iris3, frac = 3.4)
}
