% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_sig.R
\name{base_sig}
\alias{base_sig}
\title{Takes primary producer d15N signature data separated by site and returns a dataframe with two columns}
\usage{
base_sig(dat, frac = 3.4)
}
\arguments{
\item{dat}{A data frame with primary consumer (mussel) d15N data at one or more sites}

\item{frac}{Trophic discrimination factor (default set at 3.4, most commonly accepted)

The output of base_sig is a data frame with columns sitename and baseline.}
}
\value{
sitename Name of site where primary consumer (mussel) was collected

baseline A single number equal to mean primary producer d15N signature at "site"
}
\description{
Takes primary producer d15N signature data separated by site and returns a dataframe with two columns
}
\examples{
base_sig(dat = df, 3.4)
}
